requirements.txt
anthropic>=0.34.0
google-generativeai>=0.3.0
openai>=1.0.0
boto3>=1.28.0
click>=8.1.0
requests>=2.31.0
tiktoken>=0.5.0
langchain-text-splitters>=0.2.0

